ANTS-2012 + test-audit batch: dialog geometry D4 fix + 13 test/PTY fixes
ANTS-2012 (part a) — RoadmapDialog migrated onto DialogChrome::install
with resizable=true + "RoadmapDialog" sizeKey. Drops the hand-rolled
saveGeometry/restoreGeometry blob (persisted window POSITION, violating
dialogs.md D4); now persists size only (D3) + re-centers on open (D4) +
QSizeGrip (D2) via shared chrome. Orphaned Config::roadmapDialogGeometry
accessors removed. roadmap_viewer_tabs INV-12 + spec rewritten to assert
the sizeKey/no-position behaviour; dialogs.md now lists zero
non-conformers.

ANTS-1994 (finding 3) — Pty::write emits writeLost on the EAGAIN-oversize
drop path (was a silent drop, unlike the queue-full path). New INV-8 in
pty_write_eagain_queue. Finding 2 (destructor SIGHUP) investigated and
ruled not a live hazard (no parent-wide reaper); Flatpak env-policy
headline left open.

ANTS-1791 — kill -Wsign-compare in ris_preserves_callbacks (char→uint32_t).

Test-audit fixes (all single-file, deterministic-or-margin):
- ANTS-1471/1608 (dup): drop racey msleep(1100) in project_layout_scan;
  isStale comparison made deterministic via the backdated scannedAtMs.
- ANTS-1476: confirm_close extractFunctionBody → ants_test::slurpFunctionBody.
- ANTS-1477/1611: ASSERT_TRUE/EXPECT_TRUE in void/QString setup helpers
  → ADD_FAILURE (can't silently swallow open failures).
- ANTS-1601: relax brittle `int pending = 5;` source-grep to field-init.
- ANTS-1607: token_usage msleep 5→20ms (10ms-tick kernel floor).
- ANTS-1613: tool_detection cache-probe ceiling 50→200ms (CI slack).
- ANTS-1614: trim verify_trust_gate TF-2/TF-3 SHAs to canonical 64 chars.
- ANTS-1469/1470/1475: verified already-fixed-but-unflipped; closed.

Full suite 1992/1992 green.

+++ tests/features/ris_preserves_callbacks/test_ris_preserves_callbacks.cpp
@@ -110,8 +110,13 @@ TEST(RisPreservesCallbacks, Main) {
 
     // Seed a few visible cells so we can confirm RIS cleared the grid.
     feed(grid, parser, "\x1b[1;1HABC");  // CUP 1,1 then "ABC"
-    CHECK_EQ(grid.cellAt(0, 0).codepoint, 'A', "pre-RIS cell(0,0)='A'");
-    CHECK_EQ(grid.cellAt(0, 1).codepoint, 'B', "pre-RIS cell(0,1)='B'");
+    // ANTS-1791 — cast the char literal to match codepoint's uint32_t
+    // so the macro's `_a != _e` comparison isn't unsigned-vs-signed
+    // (-Wsign-compare). The counter CHECK_EQ sites above are int-vs-int.
+    CHECK_EQ(grid.cellAt(0, 0).codepoint, static_cast<uint32_t>('A'),
+             "pre-RIS cell(0,0)='A'");
+    CHECK_EQ(grid.cellAt(0, 1).codepoint, static_cast<uint32_t>('B'),
+             "pre-RIS cell(0,1)='B'");
 
     // Install an HMAC key and arm the forgery callback.
     grid.setOsc133KeyForTest(QByteArray("test-key-abc"));
